Fixing endpoints from different modules import
