+++ main.py
@@ -1,0 +1,4 @@
+from task_scheduler.app import create_app
+from task_scheduler.endpoints import api_request_tasks
+
+app = create_app()
